Move position utils to a new file
